Update signature of psi functions
I was using this fantastic article to set up local PSI tests but noticed the function signature of the `psi` module is now different than it was when the article was written. I've updated the function signatures plus I bumped the `package.json` version to make it current. Thanks again for this awesome post!