Switch the repo to the Pa11y org version
This currently points at the SpringerNature fork of Pa11y Dashboard
which doesn't work. It's our bad – `springernature/pa11y-dashboard`
used to redirect to `pa11y/dashboard` but then the company forked it.

Now it's confusing to anyone working through this post because they
potentially start working from a repo with loads of Springer Nature
specific changes, and the thing won't run 😱

Sorry if anyone's stumbled on this!